Mock data for demo
Mock data creates a user for everyone in the group. (Feel free to change any of the details just to show some examples)
 - ds@gmail.com, pw: dyllanson, privelage=ADMIN
 - ua@gmail.com, pw: uguudei, privelage=ADMIN
 - sz@gmail.com, pw: stella, privelage=MANAGER
 - mx@gmail.com, pw: meng, privelage=COLLECTOR
 - gw@gmail.com, pw: greg, privelage=COLLECTOR

Each users details are updated and the mock data is initiated as before.
collection and wantlist now displaying valid times
Pictures for collectibles and users added (collectilbes images are working)
profile_picture not showing up on front end
neither are the first_name and last_name. I've checked the backend its returning all the fields of the collector